src/requirements.txt
numpy>=2.0.0
torch>=2.5.0
transformers>=4.40.0
biopython>=1.83
tokenizers>=0.19.0
matplotlib>=3.8.0
seaborn>=0.13.0
peft>=0.10.0
